Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
httpx==0.27.0
